autorip/pyproject.toml
[tool.poetry]
name = "autorip"
version = "0.1.0"
description = ""
authors = ["DevTobias <tobi.kaerst@gmx.de>"]

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = "^3.11"
pyudev = "^0.24.1"
xmltodict = "^0.13.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
